backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
openai
python-dotenv
chromadb
python-docx
pypdf
python-multipart
